Can now start and stop programs from DialogOS. However, the EV3 brick freezes after starting a program, even if it is tracked and stopped at the end. Need to figure out why and fix it.

+++ src/main/java/elmot/javabrick/ev3/impl/Response.java
@@ -38,7 +38,7 @@ public class Response {
         return ((Number) outParameters[index]).floatValue();
     }
 
-    public int getByte(int index) {
+    public byte getByte(int index) {
         return (Byte) outParameters[index];
     }
 }
